fix: overflow error with timeout

+++ include/solana.hpp
@@ -547,7 +547,7 @@ class Connection {
    * Returns of the current Transaction has been confirmed or not
    */
   bool confirmTransaction(std::string transactionSignature,
-                          Commitment confirmLevel, uint8_t timeout=200) const;
+                          Commitment confirmLevel, uint16_t timeout=200) const;
 
   /**
    * Fetch the current statuses of a batch of signatures
